fix: metadata yaml format issue
Yaml is particular and needs spaces after colons. It can be hard to
track this down when it's in a comment block. It's even more difficult
when you are new to go and don't realize that the docstrings are pulled
from the compiled code and not the source on disk. Whoops...