docs: Provide clarity in policy titles (#140)
* docs: Provide clarity in policy titles

This commit adopts the following in an attempt to provide clarity in
policy titles and adopt some more standard RFC language.

- Critical/High rules => Must
- Medium/Low => Should

Additionally, it updates some of the policy language to be more clear.

Resolves: #139
See Also: [RFC-2119](https://www.rfc-editor.org/rfc/rfc2119)
See Also: [RFC-8174](https://www.rfc-editor.org/rfc/rfc8174)

* fix: metadata yaml format issue

Yaml is particular and needs spaces after colons. It can be hard to
track this down when it's in a comment block. It's even more difficult
when you are new to go and don't realize that the docstrings are pulled
from the compiled code and not the source on disk. Whoops...

* docs: Changed must to should

Per the conversation in [this PR](https://github.com/Legit-Labs/legitify/pull/140#issuecomment-1420547147)
I changed instanced of `must` to `should` as all of this advice is guidance
that may be followed or not at the discretion of the implementation
teams that use this tool.

---------

Co-authored-by: noamd-legit <74864790+noamd-legit@users.noreply.github.com>